requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-huggingface
langchain-pinecone
langchain-community
langchain-core
-e .